chore: added survey frequency manager
