+++ app/src/main/java/com/dwao/aliumandroidsdk/MainActivity.java
@@ -51,6 +51,6 @@ public class MainActivity extends Activity {
             }
         });
 
-        Alium.trigger(this, new SurveyParameters("firstscreen", new HashMap()));
+        Alium.trigger(this, new SurveyParameters("thirdscreen", new HashMap()));
     }
 }
